+++ src/library/index.ts
@@ -1,7 +1,5 @@
-export { ActionError } from "./error/index.tsx";
-
+export { Error } from "./error/index.tsx";
 export { createAction, createDistributedAction } from "./action/index.ts";
-
 export type {
   Pk,
   Context,
@@ -16,3 +14,4 @@ export { Broadcaster } from "./broadcast/index.tsx";
 export { useActions, useAction, useSnapshot } from "./hooks/index.ts";
 export { Operation, Op, State } from "immertation";
 export type { Box } from "immertation";
+export { use } from "./use/index.ts";